+++ src/Components/App/App.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+import Map from "../Map/Map";
+import logo from "../../Images/logo.svg";
+import "./App.css";
+
+const App = () => (
+  <div className="App">
+    <Map />
+  </div>
+);
+
+export default App;
