+++ src/Components/Map/Map.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+const Map = () => <div>This is the map</div>;
+
+export default Map;
